usb: Fix endpoint priming and buffer managing
For the I2S->USB path we can start receiving data from I2S and sending
through USB as soon as the interface is up and we keep doing that in
each device audio callback.

For the USB->I2S path when the interface is up, we instruct the USB
stack that we need to receive something from USB but we actually have
nothing to receive until the next device audio callback when the data is
ready to be picked up. So we send the data through I2S and we prime
again the endpoint for the next packet.

Signed-off-by: Carlo Caione <ccaione@baylibre.com>